+++ modules/contracts/hardhat.config.ts
@@ -99,6 +99,11 @@ const config: HardhatUserConfig = {
       chainId: 79377087078960,
       url: urlOverride || "https://kovan3.arbitrum.io/rpc",
     },
+    xdai: {
+      accounts: { mnemonic },
+      chainId: 100,
+      url: urlOverride || "http://localhost:8545",
+    },
   },
 };