+++ modules/test-runner/src/utils/channel.ts
@@ -268,19 +268,6 @@ export const withdraw = async (
   const preWithdrawMultisig = await getOnchainBalance(assetId, preWithdrawChannel.channelAddress, provider);
   const preWithdrawRecipient = await getOnchainBalance(assetId, withdrawRecipient, provider);
 
-  // // TODO: wait for these?
-  // withdrawer.on(EngineEvents.WITHDRAWAL_CREATED, (data) => {
-  //   console.log("EngineEvents.WITHDRAWAL_CREATED ===> data: ", JSON.stringify(data, null, 2));
-  // });
-
-  // withdrawer.on(EngineEvents.WITHDRAWAL_RECONCILED, (data) => {
-  //   console.log("EngineEvents.WITHDRAWAL_RECONCILED ===> data: ", JSON.stringify(data, null, 2));
-  // });
-
-  // withdrawer.on(EngineEvents.WITHDRAWAL_RESOLVED, (data) => {
-  //   console.log("EngineEvents.WITHDRAWAL_RESOLVED ===> data: ", JSON.stringify(data, null, 2));
-  // });
-
   // Perform withdrawal
   const withdrawalRes = await withdrawer.withdraw({
     publicIdentifier: withdrawer.publicIdentifier,